Release/v1.0.0

+++ Lab-Inc/src/main/java/com/labinc/Lab/Inc/services/PatientService.java
@@ -6,6 +6,9 @@ import com.labinc.Lab.Inc.entities.AllowedRoles;
 import com.labinc.Lab.Inc.entities.Patient;
 import com.labinc.Lab.Inc.entities.User;
 import com.labinc.Lab.Inc.mappers.PatientMapper;
+import com.labinc.Lab.Inc.mappers.UserMapper;
+import com.labinc.Lab.Inc.repositories.AppointmentRepository;
+import com.labinc.Lab.Inc.repositories.ExamRepository;
 import com.labinc.Lab.Inc.repositories.PatientRepository;
 import com.labinc.Lab.Inc.repositories.UserRepository;
 import com.labinc.Lab.Inc.services.exceptions.ResourceAlreadyExistsException;
@@ -18,6 +21,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Objects;
+
 
 @Service
 public class PatientService {
@@ -25,19 +30,25 @@ public class PatientService {
     private final PatientRepository patientRepository;
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    private final UserMapper userMapper;
+    private final AppointmentRepository appointmentRepository;
+    private final ExamRepository examRepository;
 
     @Autowired
-    public PatientService(PatientRepository patientRepository, UserRepository userRepository, PasswordEncoder passwordEncoder) {
+    public PatientService(PatientRepository patientRepository, UserRepository userRepository, PasswordEncoder passwordEncoder, UserMapper userMapper, AppointmentRepository appointmentRepository, ExamRepository examRepository) {
         this.patientRepository = patientRepository;
         this.userRepository = userRepository;
         this.passwordEncoder = passwordEncoder;
+        this.userMapper = userMapper;
+        this.appointmentRepository = appointmentRepository;
+        this.examRepository = examRepository;
     }
 
     @Transactional
     public PatientResponseDTO newPatient(PatientRequestDTO patientRequestDTO) {
 
         // Verifica se CPF já existe
-        if (patientRepository.existsByCpf(patientRequestDTO.getCpf())) {
+        if (patientRepository.existsByCpf(patientRequestDTO.getCpf()) || userRepository.existsByCpf(patientRequestDTO.getCpf())) {
             throw new ResourceAlreadyExistsException("O CPF já está cadastrado: " + patientRequestDTO.getCpf());
         }
 
@@ -47,7 +58,7 @@ public class PatientService {
         }
 
         // Verifica se Email já existe
-        if (patientRepository.existsByEmail(patientRequestDTO.getEmail())) {
+        if (patientRepository.existsByEmail(patientRequestDTO.getEmail()) || userRepository.existsByEmail(patientRequestDTO.getEmail())) {
             throw new ResourceAlreadyExistsException("O Email já está cadastrado: " + patientRequestDTO.getEmail());
         }
 
@@ -55,7 +66,7 @@ public class PatientService {
         User user = new User();
         user.setFullName(patientRequestDTO.getFullName());
         user.setEmail(patientRequestDTO.getEmail());
-        user.setBirthdate(patientRequestDTO.getBirthDate());
+        user.setBirthDate(patientRequestDTO.getBirthDate());
         user.setCpf(patientRequestDTO.getCpf());
         user.setPassword(passwordEncoder.encode(patientRequestDTO.getCpf())); // Configura a senha como o CPF encriptado
         user.setPasswordMasked(user.getPasswordMasked(patientRequestDTO.getCpf()));
@@ -129,7 +140,30 @@ public class PatientService {
         Patient patient = patientRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Paciente com ID: " + id + " não encontrado."));
 
+        User user = userRepository.findById(patient.getUser().getUserId())
+                .orElseThrow(() -> new ResourceNotFoundException("Usuário com ID: " + id + " não encontrado."));
+
+        if (!Objects.equals(patientRequestDTO.getCpf(), patient.getCpf())) {
+            if (patientRepository.existsByCpf(patientRequestDTO.getCpf()) || userRepository.existsByCpf(patientRequestDTO.getCpf())) {
+                throw new ResourceAlreadyExistsException("O CPF já está cadastrado: " + patientRequestDTO.getCpf());
+            }
+        }
+
+        if (!Objects.equals(patientRequestDTO.getRg(), patient.getRg())) {
+            if (patientRepository.existsByRg(patientRequestDTO.getRg())) {
+                throw new ResourceAlreadyExistsException("O RG já está cadastrado: " + patientRequestDTO.getRg());
+            }
+        }
+
+        if (!Objects.equals(patientRequestDTO.getEmail(), patient.getEmail())) {
+            if (patientRepository.existsByEmail(patientRequestDTO.getEmail()) || userRepository.existsByEmail(patientRequestDTO.getEmail())) {
+                throw new ResourceAlreadyExistsException("O Email já está cadastrado: " + patientRequestDTO.getEmail());
+            }
+        }
+
         PatientMapper.updatePatientFromDTO(patientRequestDTO, patient);
+
+        userRepository.save(userMapper.updateUserFromPatientDto(user, patientRequestDTO));
 
         Patient updatedPatient = patientRepository.save(patient);
 
@@ -137,11 +171,19 @@ public class PatientService {
     }
 
     @Transactional(propagation = Propagation.SUPPORTS)
-    public void deletePatient(Long id){
+    public void deletePatient(Long id) {
 
         if (!patientRepository.existsById(id)) {
             throw new ResourceNotFoundException("Paciente com ID: " + id + " não encontrado.");
         }
+
+        boolean hasAppointment = appointmentRepository.existsByPatient_Id(id);
+        boolean hasExam = examRepository.existsByPatient_Id(id);
+
+        if(hasAppointment || hasExam) {
+            throw new IllegalStateException("Cannot delete patient with appointments or exams");
+        }
+
         patientRepository.deleteById(id);
     }
